requirements.txt
cycler==0.12.1
numpy==2.2.2
matplotlib==3.10.0 
Pillow==11.1.0 
pyparsing==3.2.1 
python-dateutil==2.9.0
pytz==2024.2
six==1.17.0 
wheel==0.45.1 